Refactor urls to allow fine-grained control
