+++ user_management/api/tests/urls.py
@@ -2,5 +2,6 @@ from django.conf.urls import include, patterns, url
 
 
 urlpatterns = patterns('',
-    url(r'', include('user_management.api.verify_email_urls')),
+    url(r'', include('user_management.api.urls')),
+    url(r'', include('user_management.api.urls.verify_email')),
 )
