+++ user_management/api/urls/auth.py
@@ -1,0 +1,13 @@
+from django.conf.urls import patterns, url
+
+from .. import views
+
+
+urlpatterns = patterns(
+    '',
+    url(
+        regex=r'^auth/?$',
+        view=views.GetToken.as_view(),
+        name='auth',
+    ),
+)
