+++ user_management/api/urls/register.py
@@ -1,0 +1,13 @@
+from django.conf.urls import patterns, url
+
+from .. import views
+
+
+urlpatterns = patterns(
+    '',
+    url(
+        regex=r'^register/?$',
+        view=views.UserRegister.as_view(),
+        name='register',
+    ),
+)
