+++ user_management/api/urls/verify_email.py
@@ -1,11 +1,10 @@
 from django.conf.urls import patterns, url
 from django.views.decorators.csrf import csrf_exempt
 
-from .urls import urlpatterns
-from . import views
+from .. import views
 
 
-urlpatterns += patterns(
+urlpatterns = patterns(
     '',
     url(
         regex=r'^verify_email/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
